switched classes around, readapted tests and code
