+++ lib/plane.rb
@@ -1,14 +1,3 @@
-require_relative "airport.rb"
-
 class Plane
-
-  def land(plane)
-    # fail "Airport full! Plane cannot land" if
-    "Plane's landed!"
-  end
-
-  def take_off(plane)
-    "Plane took off and has left airport!"
-  end
-
+  
 end
